alx_travel_app/alx_travel_app/requirement.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
celery==5.3.4
kombu==5.3.4
drf-yasg==1.21.7
mysqlclient==2.2.0
python-dotenv==1.0.0